pyproject.toml
[project]
name = "resume"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asgiref>=3.11.1",
    "dj-database-url>=3.1.0",
    "django~=5.2.10",
    "django-cache-url>=3.4.6",
    "django-debug-toolbar>=6.2.0",
    "django-livereload-server>=0.5.1",
    "django-tinymce>=5.0.0",
    "djangorestframework>=3.16.1",
    "environs>=14.5.0",
    "pillow>=12.1.1",
    "psycopg2-binary>=2.9.11",
    "requests>=2.32.5",
    "rollbar>=1.3.0",
    "sqlparse>=0.5.5",
    "tzdata>=2025.3",
]
